fix migration and tests

+++ packages/core/src/database/migrations/z20231010145301-update-micro-credit-docs.js
@@ -2,6 +2,10 @@
 /** @type {import('sequelize-cli').Migration} */
 module.exports = {
     async up(queryInterface, Sequelize) {
+        if (process.env.NODE_ENV === 'test') {
+            return;
+        }
+
         await queryInterface.sequelize.query(`
             DELETE FROM microcredit_docs AS t1
                 WHERE id NOT IN (
